requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
django-rest-framework==0.1.0
django==4.1.2
djangorestframework==3.14.0; python_version >= '3.6'
pytz==2022.4
sqlparse==0.4.3; python_version >= '3.5'
